refactor: split schwarz preconditioner modules
