+++ crates/schwarz-precond/src/schwarz/additive.rs
@@ -1,0 +1,9 @@
+mod buffers;
+mod executor;
+mod planning;
+mod preconditioner;
+#[cfg(feature = "serde")]
+mod serde;
+
+pub use planning::{AdditiveSchwarzDiagnostics, ReductionStrategy};
+pub use preconditioner::SchwarzPreconditioner;
